Chuyển logic check thẻ chờ của TST sang class cha
