+++ dotman-plugin/src/main/java/net/minevn/dotman/providers/CardProvider.kt
@@ -4,16 +4,14 @@ import net.minevn.dotman.DotMan
 import net.minevn.dotman.DotMan.Companion.transactional
 import net.minevn.dotman.TOP_KEY_DONATE_TOTAL
 import net.minevn.dotman.TOP_KEY_POINT_FROM_CARD
-import net.minevn.dotman.card.Card
-import net.minevn.dotman.card.CardPrice
-import net.minevn.dotman.card.CardResult
-import net.minevn.dotman.card.CardType
+import net.minevn.dotman.card.*
 import net.minevn.dotman.database.LogDAO
 import net.minevn.dotman.database.PlayerDataDAO
 import net.minevn.dotman.providers.types.GameBankCP
 import net.minevn.dotman.providers.types.TheSieuTocCP
 import net.minevn.dotman.utils.Utils.Companion.runNotSync
 import net.minevn.dotman.utils.Utils.Companion.send
+import net.minevn.dotman.utils.Utils.Companion.severe
 import net.minevn.dotman.utils.Utils.Companion.warning
 import net.minevn.libs.bukkit.chat.ChatListener
 import net.minevn.libs.bukkit.sendMessages
@@ -195,5 +193,38 @@ abstract class CardProvider {
     }
 
     protected open fun updateStatus() {
+    }
+
+    protected open fun CardWaiting.isProcessed() = false
+
+    protected fun checkWaitingCards() {
+        val players = Bukkit.getOnlinePlayers().associateBy { it.uniqueId.toString() }
+        val uuids = players.keys.toTypedArray()
+        val server = main.config.server
+        val cardLogger = LogDAO.getInstance()
+        val lang = main.language
+
+        cardLogger.getWaitingCards(uuids, server)
+            ?.filter { it.isProcessed() }
+            ?.forEach {
+                val player = players[it.uuid]
+                if (player?.isOnline != true) return@forEach
+
+                cardLogger.stopWaiting(it.id, it.isSuccess)
+
+                if (it.isSuccess) {
+                    try {
+                        onChargeSuccess(player, it.toCard())
+                    } catch (e: Exception) {
+                        player.sendMessages(lang.cardChargedError)
+                        e.severe("Error onChargeSuccess: Player ${player.name}, ${it.toCard()}")
+                    }
+                } else {
+                    val message = it.message ?: lang.errorUnknown
+                    player.sendMessages(lang.cardChargedFailed.map { str ->
+                        str.replace("%ERROR%", message)
+                    })
+                }
+            }
     }
 }
